function_modules/sprocket-encode/requirements.txt
opencv-python
redis
ffmpeg-python

zlib-state
pilgram
Pillow==6.2.2
paramiko==2.11.0
pycparser==2.21
PyNaCl==1.5.0
pysftp==0.2.9